Reduced redunfant duplicated code by centralizing tasks in a RowFactory
