+++ src/main/java/cs2103/v15_1j/jimjim/ui/FloatingTaskPaneController.java
@@ -20,6 +20,7 @@ public class FloatingTaskPaneController {
 	private ScrollPane floatingTaskScrollPane;
 
 	private MainViewController con;
+	private TaskEventRowFactory rowFactory;
 
 	private DataLists masterList;
 	private DataLists displayList;
@@ -45,7 +46,7 @@ public class FloatingTaskPaneController {
 
 	private void initialize(){
 		setUpFloatingTaskPane();
-		showFloatingTasks();
+		//showFloatingTasks();
 	}
 
 	private void setUpFloatingTaskPane(){
@@ -60,105 +61,13 @@ public class FloatingTaskPaneController {
 		floatingTaskScrollPane.setFocusTraversable(false);
 		floatingTaskScrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
 		floatingTaskScrollPane.getStyleClass().add("pane");
+		
+		rowFactory = new TaskEventRowFactory(masterList, displayList, floatingTaskGridPane, new Integer(0));
 	}
 
 	public void refreshData(){
-		showFloatingTasks();
+		//showFloatingTasks();
 	}
 
-	private void showFloatingTasks(){
-		floatingTaskGridPane.getChildren().clear();
-		int counter = 0;
-		int rowNo = 0;
-		hasCompleted = false;
-
-		Label titleLabel = new Label("Floating Tasks");
-		titleLabel.getStyleClass().add("header");
-		floatingTaskGridPane.add(titleLabel, 0, rowNo, 4, 1);
-
-		for(FloatingTask t: masterList.getFloatingTasksList()){
-			if(!t.getCompleted()){
-				rowNo++;
-				addFloatingTaskToPane(t);
-			}
-			else {
-				hasCompleted = true;
-			}
-		}
-
-		if(showCompleted){
-			for(FloatingTask t: masterList.getFloatingTasksList()){
-				if(t.getCompleted()){
-					rowNo++;
-					addFloatingTaskToPane(t);
-				}
-			}
-		}
-
-		if(rowNo == 0){
-			rowNo++;
-			Label noLabel = new Label("There are no outstanding floating tasks.");
-			noLabel.getStyleClass().add("red-label");
-			floatingTaskGridPane.add(noLabel, 0, rowNo, 1, 1);
-		}
-
-		rowNo++;
-
-		if(hasCompleted && !showCompleted){
-			floatingTaskGridPane.getChildren().remove(titleLabel);
-			floatingTaskGridPane.add(titleLabel, 0, 0, 3, 1);
-			JFXButton showCompletedBtn = new JFXButton("Show Completed");
-			showCompletedBtn.getStyleClass().add("button-raised");
-			showCompletedBtn.setOnAction(event -> toggleShowCompleted());
-			GridPane.setHalignment(showCompletedBtn, HPos.CENTER);
-			floatingTaskGridPane.add(showCompletedBtn, 2, 0, 1, 1);
-		}
-		else if (hasCompleted && showCompleted){
-			floatingTaskGridPane.getChildren().remove(titleLabel);
-			floatingTaskGridPane.add(titleLabel, 0, 0, 3, 1);
-			JFXButton hideCompletedBtn = new JFXButton("Hide Completed");
-			hideCompletedBtn.getStyleClass().add("button-raised");
-			hideCompletedBtn.setOnAction(event -> toggleShowCompleted());
-			GridPane.setHalignment(hideCompletedBtn, HPos.CENTER);
-			floatingTaskGridPane.add(hideCompletedBtn, 2, 0, 1, 1);
-		}
-
-	}
-
-	private void addFloatingTaskToPane(FloatingTask t){
-		displayList.addWithoutSorting(t);
-		
-		JFXCheckBox cb = new JFXCheckBox();
-		cb.getStyleClass().add("custom-jfx-check-box");
-		cb.selectedProperty().bindBidirectional(t.completedProperty());
-		cb.setDisable(true);
-		floatingTaskGridPane.addColumn(0, cb);
-		cb.setOnMouseClicked(event -> showFloatingTasks());
-
-		int id = masterList.indexOf(t) + 1;
-		Label idLabel = new Label("[F"+displayList.size('f')+"]");
-		floatingTaskGridPane.addColumn(1, idLabel);
-
-		Label taskLabel = new Label();
-		taskLabel.textProperty().bindBidirectional(t.taskNameProperty());
-		taskLabel.setWrapText(true);
-		taskLabel.setPrefWidth(TASK_LABEL_WIDTH);
-
-		if(!t.getCompleted()){
-			idLabel.getStyleClass().add("id-label");
-			taskLabel.getStyleClass().add("task-label");
-		}
-		else {
-			idLabel.getStyleClass().add("completed-task-label");
-			taskLabel.getStyleClass().add("completed-task-label");
-		}
-
-		floatingTaskGridPane.addColumn(2, taskLabel);
-	}
-
-	private void toggleShowCompleted(){
-		showCompleted = !showCompleted;
-		con.updateData();
-		con.focusCommandBar();
-	}
+	
 }
