+++ src/main/java/cs2103/v15_1j/jimjim/ui/SearchPaneController.java
@@ -22,10 +22,13 @@ public class SearchPaneController {
 	private ScrollPane searchScrollPane;
 
 	private MainViewController con;
+	private TaskEventRowFactory rowFactory;
 
 	private DataLists masterList;
 	private DataLists searchResultsList;
 	private DataLists displayList;
+	
+	private Integer rowNo;
 
 	private final double COLUMN_WIDTH = 500.0;
 	private final double NAME_LABEL_WIDTH = 250.0;
@@ -46,7 +49,7 @@ public class SearchPaneController {
 
 	private void initialize(){
 		setUpSearchPane();
-		showSearchResults();
+		rowFactory.showSearchResults();
 	}
 
 	private void setUpSearchPane(){
@@ -61,131 +64,12 @@ public class SearchPaneController {
 		searchScrollPane.setFocusTraversable(false);
 		searchScrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
 		searchScrollPane.getStyleClass().add("pane");
+		
+		rowFactory = new TaskEventRowFactory(searchResultsList, displayList, searchGridPane, rowNo);
 	}
 
 	public void refreshData(){
-		showSearchResults();
-	}
-
-	private void showSearchResults(){
-		searchGridPane.getChildren().clear();
-		int rowNo = 0;
-
-		Label titleLabel = new Label("Search Results");
-		titleLabel.getStyleClass().add("header");
-		searchGridPane.add(titleLabel, 0, rowNo, 4, 1);
-
-		int counter = 0;
-		
-		for(Event event: searchResultsList.getEventsList()){
-			counter++;
-			displayList.addWithoutSorting(event);
-			
-			Circle dot = new Circle(3.0, Color.RED);
-			GridPane.setHalignment(dot, HPos.CENTER);
-			searchGridPane.addColumn(0, dot);
-
-			int id = displayList.size('e');
-			Label idLabel = new Label("[E"+id+"]");
-			idLabel.getStyleClass().add("id-label");
-			searchGridPane.addColumn(1, idLabel);
-
-			Label eventLabel = new Label();
-			eventLabel.getStyleClass().add("event-label");
-			eventLabel.textProperty().bindBidirectional(event.taskNameProperty());
-			eventLabel.setTextAlignment(TextAlignment.LEFT);
-			eventLabel.setWrapText(true);
-			eventLabel.setPrefWidth(NAME_LABEL_WIDTH);
-			searchGridPane.addColumn(2, eventLabel);
-
-			Label dateLabel = new Label(event.toString());
-			dateLabel.getStyleClass().add("event-label");
-			dateLabel.setTextAlignment(TextAlignment.RIGHT);
-			dateLabel.setWrapText(true);
-			dateLabel.setPrefWidth(DATE_LABEL_WIDTH);
-			searchGridPane.addColumn(3, dateLabel);
-		}
-
-		for(DeadlineTask task: searchResultsList.getDeadlineTasksList()){
-			counter++;
-			displayList.addWithoutSorting(task);
-			
-			JFXCheckBox cb = new JFXCheckBox();
-			cb.getStyleClass().add("custom-jfx-check-box");
-			cb.selectedProperty().bindBidirectional(task.completedProperty());
-			cb.setDisable(true);
-			GridPane.setHalignment(cb, HPos.CENTER);
-			searchGridPane.addColumn(0, cb);
-
-			int id = displayList.size('d');
-			Label idLabel = new Label("[D"+id+"]");
-			searchGridPane.addColumn(1, idLabel);
-
-			Label taskLabel = new Label();
-			taskLabel.textProperty().bindBidirectional(task.taskNameProperty());
-			taskLabel.setWrapText(true);
-			taskLabel.setPrefWidth(NAME_LABEL_WIDTH);
-			searchGridPane.addColumn(2, taskLabel);
-
-			DateTimeFormatter dateFmt = DateTimeFormatter.ofPattern("dd MMM h:mm a");
-			Label dateTimeLabel = new Label(task.getDateTime().format(dateFmt));
-			dateTimeLabel.setTextAlignment(TextAlignment.RIGHT);
-			dateTimeLabel.setWrapText(true);
-			dateTimeLabel.setPrefWidth(DATE_LABEL_WIDTH);
-
-			if(!task.getCompleted()){
-				idLabel.getStyleClass().add("id-label");
-				taskLabel.getStyleClass().add("task-label");
-				dateTimeLabel.getStyleClass().add("task-label");
-			}
-			else {
-				idLabel.getStyleClass().add("completed-task-label");
-				taskLabel.getStyleClass().add("completed-task-label");
-				dateTimeLabel.getStyleClass().add("completed-task-label");
-			}
-			
-			searchGridPane.addColumn(3, dateTimeLabel);
-		}
-
-		for(FloatingTask task: searchResultsList.getFloatingTasksList()){
-			counter++;
-			displayList.addWithoutSorting(task);
-			
-			JFXCheckBox cb = new JFXCheckBox();
-			cb.getStyleClass().add("custom-jfx-check-box");
-			cb.setDisable(true);
-			cb.selectedProperty().bindBidirectional(task.completedProperty());
-			searchGridPane.addColumn(0, cb);
-			cb.setOnMouseClicked(event -> showSearchResults());
-
-			int id = displayList.size('f');
-			Label idLabel = new Label("[F"+id+"]");
-			searchGridPane.addColumn(1, idLabel);
-
-			Label taskLabel = new Label();
-			taskLabel.textProperty().bindBidirectional(task.taskNameProperty());
-			taskLabel.setWrapText(true);
-			taskLabel.setPrefWidth(FLOATING_NAME_LABEL_WIDTH);
-			searchGridPane.add(taskLabel, 2, counter, 2, 1);
-
-			if(!task.getCompleted()){
-				idLabel.getStyleClass().add("id-label");
-				taskLabel.getStyleClass().add("task-label");
-			}
-			else {
-				idLabel.getStyleClass().add("completed-task-label");
-				taskLabel.getStyleClass().add("completed-task-label");
-			}
-
-			Label blankLabel = new Label();
-			searchGridPane.addColumn(3, blankLabel);
-		}
-		
-		if(counter == 0){
-			Label emptyLabel = new Label("No results match your search term");
-			emptyLabel.getStyleClass().add("red-label");
-			searchGridPane.add(emptyLabel, 0, 1, 4, 1);
-		}
+		rowFactory.showSearchResults();
 	}
 
 }
