+++ src/main/resources/css/ui.css
@@ -29,7 +29,6 @@
 .overdue-label{
 	-fx-text-fill: #ff3333;
     -fx-font-family: "Roboto Thin";
-    -fx-font-size: 16pt;
 }
 
 .no-overdue-label{
